refactor: use `expect class`
