+++ composeApp/src/commonMain/kotlin/paige/navic/ui/screens/album/components/Item.kt
@@ -18,7 +18,7 @@ import navic.composeapp.generated.resources.notice_download_started
 import org.koin.compose.koinInject
 import paige.navic.LocalNavStack
 import paige.navic.data.database.entities.DownloadStatus
-import paige.navic.domain.manager.downloads.DownloadManager
+import paige.navic.domain.manager.DownloadManager
 import paige.navic.domain.manager.SnackBarManager
 import paige.navic.domain.models.DomainAlbum
 import paige.navic.ui.components.layouts.ArtGridItem
