+++ composeApp/src/commonMain/kotlin/paige/navic/ui/screens/queue/viewmodels/QueueViewModel.kt
@@ -3,7 +3,7 @@ package paige.navic.ui.screens.queue.viewmodels
 import androidx.compose.foundation.lazy.LazyListState
 import androidx.lifecycle.ViewModel
 import paige.navic.domain.manager.ConnectivityManager
-import paige.navic.domain.manager.downloads.DownloadManager
+import paige.navic.domain.manager.DownloadManager
 
 class QueueViewModel(
 	connectivityManager: ConnectivityManager,
