+++ composeApp/src/iosMain/kotlin/paige/navic/domain/manager/base/BaseDownloadManager.ios.kt
@@ -1,4 +1,4 @@
-package paige.navic.domain.manager
+package paige.navic.domain.manager.base
 
 import io.ktor.client.HttpClient
 import io.ktor.client.plugins.onDownload
@@ -6,20 +6,20 @@ import io.ktor.client.request.header
 import io.ktor.client.request.prepareRequest
 import io.ktor.client.statement.bodyAsChannel
 import io.ktor.http.HttpMethod
-import paige.navic.domain.manager.downloads.BaseDownloadManager
+import paige.navic.domain.manager.StorageManager
 import paige.navic.domain.models.DomainSong
 
-class IosBaseDownloadManager(
+actual class BaseDownloadManager(
 	private val storageManager: StorageManager
-) : BaseDownloadManager {
+) {
 	private val client = HttpClient()
 
-	override suspend fun downloadAudio(
-		song: DomainSong,
-		url: String,
-		extension: String,
-		headers: Map<String, String>,
-		onProgress: suspend (Float) -> Unit
+	actual suspend fun downloadAudio(
+        song: DomainSong,
+        url: String,
+        extension: String,
+        headers: Map<String, String>,
+        onProgress: suspend (Float) -> Unit
 	): String {
 		val request = client.prepareRequest(url) {
 			method = HttpMethod.Get
@@ -38,7 +38,7 @@ class IosBaseDownloadManager(
 			finalPath = storageManager.getDownloadPath(song.id, extension)
 			storageManager.saveFile(finalPath, response.bodyAsChannel())
 		}
-		
+
 		return finalPath
 	}
 }
